[RNE Rewrite] feat(demo): merge OCR + document screens into one progressive OCR screen
