+++ apps/computer-vision/app/_layout.tsx
@@ -84,13 +84,6 @@ export default function Layout() {
         }}
       />
       <Drawer.Screen
-        name="document/index"
-        options={{
-          drawerLabel: 'Document Pipeline',
-          title: 'Document Pipeline',
-        }}
-      />
-      <Drawer.Screen
         name="inspect/index"
         options={{
           drawerLabel: 'Model Inspector',
