+++ apps/computer-vision/app/index.tsx
@@ -41,9 +41,6 @@ export default function Home() {
         <TouchableOpacity style={styles.button} onPress={() => router.navigate('ocr/')}>
           <Text style={styles.buttonText}>OCR</Text>
         </TouchableOpacity>
-        <TouchableOpacity style={styles.button} onPress={() => router.navigate('document/')}>
-          <Text style={styles.buttonText}>Document Pipeline</Text>
-        </TouchableOpacity>
         <TouchableOpacity style={styles.button} onPress={() => router.navigate('inspect/')}>
           <Text style={styles.buttonText}>Model Inspector</Text>
         </TouchableOpacity>
