+++ apps/computer-vision/app/ocr/index.tsx
@@ -1,8 +1,14 @@
 import React, { useMemo, useState } from 'react';
 import { View, Text, StyleSheet, ScrollView, Platform, Switch } from 'react-native';
 import { commonStyles, ColorPalette } from '../../theme';
-import { useImage } from '@shopify/react-native-skia';
-import { useOcr, models, type OcrDetection } from 'react-native-executorch';
+import { useImage, Skia, ColorType, AlphaType, type SkImage } from '@shopify/react-native-skia';
+import {
+  useOcr,
+  models,
+  type OcrDetection,
+  type DocumentBlock,
+  type OcrModel,
+} from 'react-native-executorch';
 import ScreenWrapper from '../../components/ScreenWrapper';
 import { getImage } from '../../utils';
 import { ModelPicker, type ModelOption } from '../../components/ModelPicker';
@@ -12,95 +18,183 @@ import { Button } from '../../components/Button';
 
 const PREVIEW_HEIGHT = 280;
 
-// Hosted PTEs — downloaded + cached on-device from Hugging Face by `useOcr`.
-// Backends per platform: XNNPACK runs everywhere, Vulkan on Android, CoreML on iOS.
-const ALL_MODELS = [
+// Hosted base OCR models, downloaded + cached on-device by `useOcr`. Backends per
+// platform: XNNPACK runs everywhere, Vulkan on Android, CoreML on iOS. The layout
+// and document-helper models (added by the toggles below) use the same backend.
+type BackendKey = 'XNNPACK' | 'VULKAN' | 'COREML';
+const ALL_MODELS: { label: string; backend: BackendKey; base: OcrModel; platforms: string[] }[] = [
   {
     label: 'PaddleOCR (XNNPACK)',
-    config: models.ocr.PADDLE.PPOCRV6_SMALL.XNNPACK,
+    backend: 'XNNPACK',
+    base: models.ocr.PADDLE.PPOCRV6_SMALL.XNNPACK,
     platforms: ['ios', 'android'],
   },
   {
     label: 'PaddleOCR (Vulkan)',
-    config: models.ocr.PADDLE.PPOCRV6_SMALL.VULKAN,
+    backend: 'VULKAN',
+    base: models.ocr.PADDLE.PPOCRV6_SMALL.VULKAN,
     platforms: ['android'],
   },
   {
     label: 'PaddleOCR (CoreML)',
-    config: models.ocr.PADDLE.PPOCRV6_SMALL.COREML,
+    backend: 'COREML',
+    base: models.ocr.PADDLE.PPOCRV6_SMALL.COREML,
     platforms: ['ios'],
   },
   {
     label: 'EasyOCR English (XNNPACK)',
-    config: models.ocr.EASYOCR.ENGLISH.XNNPACK,
+    backend: 'XNNPACK',
+    base: models.ocr.EASYOCR.ENGLISH.XNNPACK,
     platforms: ['ios', 'android'],
   },
   {
     label: 'EasyOCR English (Vulkan)',
-    config: models.ocr.EASYOCR.ENGLISH.VULKAN,
+    backend: 'VULKAN',
+    base: models.ocr.EASYOCR.ENGLISH.VULKAN,
     platforms: ['android'],
   },
   {
     label: 'EasyOCR English (CoreML)',
-    config: models.ocr.EASYOCR.ENGLISH.COREML,
+    backend: 'COREML',
+    base: models.ocr.EASYOCR.ENGLISH.COREML,
     platforms: ['ios'],
   },
 ];
 
 const OCR_MODELS = ALL_MODELS.filter((m) => m.platforms.includes(Platform.OS));
-
 const MODEL_OPTIONS: ModelOption[] = OCR_MODELS.map((m, i) => ({ label: m.label, value: i }));
+
+type Cell = { text: string; colspan: number };
+
+// Parse the filled SLANet structure HTML into rows of cells for rendering.
+function parseTable(html: string): Cell[][] {
+  const rows: Cell[][] = [];
+  const trRe = /<tr>([\s\S]*?)<\/tr>/g;
+  let tr: RegExpExecArray | null;
+  while ((tr = trRe.exec(html))) {
+    const cells: Cell[] = [];
+    const tdRe = /<td([^>]*)>([\s\S]*?)<\/td>/g;
+    let td: RegExpExecArray | null;
+    while ((td = tdRe.exec(tr[1]!))) {
+      cells.push({
+        text: td[2] ?? '',
+        colspan: Number(/colspan="(\d+)"/.exec(td[1] ?? '')?.[1] ?? 1),
+      });
+    }
+    rows.push(cells);
+  }
+  return rows;
+}
+
+function TableView({ html }: { html: string }) {
+  const rows = parseTable(html);
+  if (rows.length === 0) {
+    return <Text style={styles.blockText}>{html}</Text>;
+  }
+  return (
+    <ScrollView horizontal showsHorizontalScrollIndicator={false}>
+      <View style={styles.table}>
+        {rows.map((cells, r) => (
+          <View key={r} style={styles.tr}>
+            {cells.map((c, i) => (
+              <View key={i} style={[styles.td, { width: 110 * c.colspan }]}>
+                <Text style={styles.tdText}>{c.text}</Text>
+              </View>
+            ))}
+          </View>
+        ))}
+      </View>
+    </ScrollView>
+  );
+}
 
 function OCRContent() {
   const [selectedIdx, setSelectedIdx] = useState(0);
   const [vertical, setVertical] = useState(false);
+  const [layoutOn, setLayoutOn] = useState(false);
+  const [documentOn, setDocumentOn] = useState(false);
+  const [orientation, setOrientation] = useState(true);
+  // Off by default: dewarp corrects photographed, physically-warped pages; on a
+  // flat screenshot it has nothing to fix and visibly distorts clean text.
+  const [dewarp, setDewarp] = useState(false);
   const [imageUri, setImageUri] = useState<string | null>(null);
   const [isProcessing, setIsProcessing] = useState(false);
-  const [results, setResults] = useState<OcrDetection[]>([]);
+  const [detections, setDetections] = useState<OcrDetection[]>([]);
+  const [blocks, setBlocks] = useState<DocumentBlock<string>[]>([]);
+  // The frame the result boxes are relative to (orientation/dewarp may move it
+  // away from the original), so the overlay lines up.
+  const [processed, setProcessed] = useState<SkImage | null>(null);
   const [wallMs, setWallMs] = useState<number | null>(null);
   const [error, setError] = useState<string | null>(null);
 
   const selected = OCR_MODELS[selectedIdx]!;
-
   const skiaImage = useImage(imageUri, (err) => setError(err.message || String(err)));
 
-  const { isReady, downloadProgress, error: loadError, runOcr } = useOcr(selected.config);
+  // `useOcr` downloads + caches each enabled model. layout / document helpers are
+  // added only when toggled; orientation/dewarp are per-run (no reload to toggle).
+  const config = {
+    ...selected.base,
+    ...(layoutOn ? { layout: models.layoutDetection.PP_DOCLAYOUT[selected.backend] } : {}),
+    ...(documentOn ? { documentModels: models.documentModels.PP_HELPERS[selected.backend] } : {}),
+  };
 
-  const handlePickImage = async (useCamera: boolean) => {
+  const { isReady, downloadProgress, error: loadError, runOcr } = useOcr<string>(config);
+
+  const resetResults = () => {
+    setDetections([]);
+    setBlocks([]);
+    setProcessed(null);
+    setWallMs(null);
+  };
+
+  const handlePick = async (useCamera: boolean) => {
     setError(null);
     try {
       const uri = await getImage(useCamera);
       if (uri) {
         setImageUri(uri);
-        setResults([]);
-        setWallMs(null);
+        resetResults();
       }
     } catch (e: any) {
       setError(e.message || String(e));
     }
   };
 
-  const runRecognition = async () => {
+  const run = async () => {
     if (!skiaImage || !runOcr) return;
     setIsProcessing(true);
     setError(null);
     try {
       const pixels = skiaImage.readPixels();
-      if (!(pixels instanceof Uint8Array)) {
-        throw new Error('Expected Uint8Array from readPixels');
-      }
-      const buffer = {
-        data: pixels,
-        width: skiaImage.width(),
-        height: skiaImage.height(),
-        format: 'rgba' as const,
-        layout: 'hwc' as const,
-      };
+      if (!(pixels instanceof Uint8Array)) throw new Error('Expected Uint8Array from readPixels');
       const start = Date.now();
-      // `vertical` is a per-run option now — toggling it needs no model reload.
-      const output = await runOcr(buffer, { vertical });
+      const out = await runOcr(
+        {
+          data: pixels,
+          width: skiaImage.width(),
+          height: skiaImage.height(),
+          format: 'rgba' as const,
+          layout: 'hwc' as const,
+        },
+        { vertical, orientation: documentOn && orientation, dewarp: documentOn && dewarp }
+      );
       setWallMs(Date.now() - start);
-      setResults(output.detections);
+      setDetections(out.detections);
+      setBlocks(out.blocks);
+      // Show the frame the boxes are relative to (orientation/dewarp may have
+      // rotated/warped it) so the overlaid boxes align.
+      const frame = out.image;
+      const frameImage = Skia.Image.MakeImage(
+        {
+          width: frame.width,
+          height: frame.height,
+          colorType: ColorType.RGBA_8888,
+          alphaType: AlphaType.Premul,
+        },
+        Skia.Data.fromBytes(frame.data),
+        frame.width * 4
+      );
+      setProcessed(frameImage ?? null);
     } catch (e: any) {
       setError(e.message || String(e));
     } finally {
@@ -109,7 +203,9 @@ function OCRContent() {
   };
 
   const activeError = loadError ? String(loadError) : error;
-  const boxes = useMemo(() => results.map((r) => r.quad), [results]);
+  // Overlay the precise per-line quads (always present, in the corrected frame).
+  const boxes = useMemo(() => detections.map((d) => d.quad), [detections]);
+  const showBlocks = layoutOn && blocks.length > 0;
 
   return (
     <ScrollView
@@ -117,7 +213,8 @@ function OCRContent() {
       contentContainerStyle={commonStyles.contentContainer}
     >
       <Text style={commonStyles.description}>
-        Upload or capture an image to detect and recognize text on-device.
+        Detect and recognize text on-device. Turn on Layout to group text into reading-ordered
+        blocks, and Document helpers for orientation, table structure and dewarp.
       </Text>
 
       <ModelPicker
@@ -126,21 +223,46 @@ function OCRContent() {
         selectedValue={selectedIdx}
         onValueChange={(idx) => {
           setSelectedIdx(idx);
-          setResults([]);
-          setWallMs(null);
+          resetResults();
           setError(null);
         }}
       />
 
-      <View style={styles.toggleRow}>
-        <View style={styles.toggleText}>
-          <Text style={styles.toggleLabel}>Vertical text</Text>
-          <Text style={styles.toggleHint}>
-            Read upright stacked columns (character-under-character)
-          </Text>
-        </View>
-        <Switch value={vertical} onValueChange={setVertical} />
-      </View>
+      <Toggle
+        label="Vertical text"
+        value={vertical}
+        onChange={setVertical}
+        hint="read upright stacked columns"
+      />
+      <Toggle
+        label="Layout (blocks)"
+        value={layoutOn}
+        onChange={setLayoutOn}
+        hint="group into reading-ordered regions"
+      />
+      <Toggle
+        label="Document helpers"
+        value={documentOn}
+        onChange={setDocumentOn}
+        hint="orientation, table structure, dewarp"
+      />
+      {documentOn && (
+        <>
+          <Toggle
+            label="Correct orientation"
+            value={orientation}
+            onChange={setOrientation}
+            indent
+          />
+          <Toggle
+            label="Dewarp"
+            value={dewarp}
+            onChange={setDewarp}
+            indent
+            hint="warped photos only"
+          />
+        </>
+      )}
 
       <ModelStatus
         isReady={isReady}
@@ -150,21 +272,20 @@ function OCRContent() {
       />
 
       <ImageViewport
-        skiaImage={skiaImage}
+        skiaImage={processed ?? skiaImage}
         height={PREVIEW_HEIGHT}
         boxes={boxes}
-        onPressPlaceholder={() => handlePickImage(false)}
+        onPressPlaceholder={() => handlePick(false)}
       />
 
       <View style={commonStyles.buttonRow}>
-        <Button title="Gallery" onPress={() => handlePickImage(false)} variant="secondary" />
-        <Button title="Camera" onPress={() => handlePickImage(true)} variant="secondary" />
+        <Button title="Gallery" onPress={() => handlePick(false)} variant="secondary" />
+        <Button title="Camera" onPress={() => handlePick(true)} variant="secondary" />
       </View>
-
       <View style={commonStyles.buttonRow}>
         <Button
           title="Run OCR"
-          onPress={runRecognition}
+          onPress={run}
           disabled={!skiaImage || !isReady || isProcessing}
           loading={isProcessing}
         />
@@ -182,29 +303,68 @@ function OCRContent() {
               <Text style={styles.tileLabel}>Wall time</Text>
             </View>
             <View style={styles.tile}>
-              <Text style={styles.tileValue}>{results.length}</Text>
-              <Text style={styles.tileLabel}>Regions read</Text>
+              <Text style={styles.tileValue}>{showBlocks ? blocks.length : detections.length}</Text>
+              <Text style={styles.tileLabel}>{showBlocks ? 'Blocks' : 'Regions read'}</Text>
             </View>
           </View>
         </View>
       )}
 
-      {results.length > 0 && (
-        <View style={styles.resultsContainer}>
-          <Text style={styles.resultsTitle}>Detected text ({results.length})</Text>
-          {results.map((res, idx) => (
-            <View key={idx} style={styles.resultRow}>
-              <Text style={styles.resultLabel} numberOfLines={1}>
-                {res.text}
+      {showBlocks ? (
+        <View style={styles.results}>
+          <Text style={styles.resultsTitle}>Blocks ({blocks.length})</Text>
+          {blocks.map((b, i) => (
+            <View key={i} style={styles.block}>
+              <Text style={styles.regionType}>
+                {b.regionType}
+                {b.isTable ? '  · table' : ''}
               </Text>
-              <View style={styles.resultMeta}>
-                <Text style={styles.resultConfidence}>{Math.round(res.confidence * 100)}%</Text>
-              </View>
+              {b.isTable && b.tableHtml ? (
+                <TableView html={b.tableHtml} />
+              ) : (
+                <Text style={styles.blockText}>{b.text}</Text>
+              )}
             </View>
           ))}
         </View>
-      )}
+      ) : detections.length > 0 ? (
+        <View style={styles.results}>
+          <Text style={styles.resultsTitle}>Detected text ({detections.length})</Text>
+          {detections.map((d, i) => (
+            <View key={i} style={styles.resultRow}>
+              <Text style={styles.resultLabel} numberOfLines={1}>
+                {d.text}
+              </Text>
+              <Text style={styles.resultConfidence}>{Math.round(d.confidence * 100)}%</Text>
+            </View>
+          ))}
+        </View>
+      ) : null}
     </ScrollView>
+  );
+}
+
+function Toggle({
+  label,
+  value,
+  onChange,
+  hint,
+  indent,
+}: {
+  label: string;
+  value: boolean;
+  onChange: (v: boolean) => void;
+  hint?: string;
+  indent?: boolean;
+}) {
+  return (
+    <View style={[styles.toggleRow, indent && styles.toggleIndent]}>
+      <View style={styles.toggleText}>
+        <Text style={styles.toggleLabel}>{label}</Text>
+        {hint ? <Text style={styles.toggleHint}>{hint}</Text> : null}
+      </View>
+      <Switch value={value} onValueChange={onChange} />
+    </View>
   );
 }
 
@@ -222,8 +382,9 @@ const styles = StyleSheet.create({
     alignItems: 'center',
     justifyContent: 'space-between',
     width: '100%',
-    marginBottom: 12,
+    marginBottom: 8,
   },
+  toggleIndent: { paddingLeft: 16 },
   toggleText: { flex: 1, marginRight: 12 },
   toggleLabel: { fontSize: 15, fontWeight: '600', color: ColorPalette.strongPrimary },
   toggleHint: { fontSize: 12, color: '#868e96', marginTop: 2 },
@@ -232,7 +393,7 @@ const styles = StyleSheet.create({
     backgroundColor: '#fff',
     borderRadius: 12,
     padding: 16,
-    marginBottom: 16,
+    marginVertical: 16,
     borderWidth: 1,
     borderColor: '#e9ecef',
   },
@@ -244,11 +405,7 @@ const styles = StyleSheet.create({
     textTransform: 'uppercase',
     marginBottom: 12,
   },
-  statTiles: {
-    flexDirection: 'row',
-    gap: 12,
-    marginBottom: 12,
-  },
+  statTiles: { flexDirection: 'row', gap: 12 },
   tile: {
     flex: 1,
     backgroundColor: '#f2f4ff',
@@ -256,16 +413,10 @@ const styles = StyleSheet.create({
     paddingVertical: 12,
     paddingHorizontal: 14,
   },
-  tileValue: {
-    fontSize: 24,
-    fontWeight: '800',
-    color: '#001A72',
-    fontVariant: ['tabular-nums'],
-  },
+  tileValue: { fontSize: 24, fontWeight: '800', color: '#001A72', fontVariant: ['tabular-nums'] },
   tileUnit: { fontSize: 14, fontWeight: '600', color: '#6b73a3' },
   tileLabel: { fontSize: 11, color: '#868e96', marginTop: 4 },
-  resultMeta: { flexDirection: 'row', alignItems: 'center' },
-  resultsContainer: {
+  results: {
     width: '100%',
     backgroundColor: '#fff',
     borderRadius: 12,
@@ -282,19 +433,30 @@ const styles = StyleSheet.create({
   resultRow: {
     flexDirection: 'row',
     justifyContent: 'space-between',
+    alignItems: 'center',
     paddingVertical: 8,
     borderBottomWidth: 1,
     borderBottomColor: '#f1f3f5',
   },
-  resultLabel: {
-    fontSize: 14,
-    color: '#333',
-    flex: 1,
-    marginRight: 8,
-  },
-  resultConfidence: {
-    fontSize: 14,
-    fontWeight: '600',
+  resultLabel: { fontSize: 14, color: '#333', flex: 1, marginRight: 8 },
+  resultConfidence: { fontSize: 14, fontWeight: '600', color: '#2b8a3e' },
+  block: { paddingVertical: 8, borderBottomWidth: 1, borderBottomColor: '#f1f3f5' },
+  regionType: {
+    fontSize: 12,
+    fontWeight: '700',
     color: '#2b8a3e',
+    textTransform: 'uppercase',
+    marginBottom: 4,
   },
+  blockText: { fontSize: 14, color: '#333' },
+  table: { borderWidth: 1, borderColor: '#ced4da', borderRadius: 4, overflow: 'hidden' },
+  tr: { flexDirection: 'row' },
+  td: {
+    borderWidth: StyleSheet.hairlineWidth,
+    borderColor: '#ced4da',
+    paddingHorizontal: 6,
+    paddingVertical: 4,
+    minWidth: 24,
+  },
+  tdText: { fontSize: 13, color: '#333' },
 });
